[feat] #7 domain model 재정비
