+++ src/main/java/com/spring/guideance/post/domain/Likes.java
@@ -1,0 +1,26 @@
+package com.spring.guideance.post.domain;
+
+import com.spring.guideance.user.domain.User;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Entity
+@Getter @Setter
+public class Likes {
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "likes_id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "article_id")
+    private Article article;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    private LocalDateTime createdAt;
+}
