+++ src/main/java/com/spring/guideance/tag/domain/ArticleTag.java
@@ -1,12 +1,10 @@
-package com.spring.guideance.tag;
+package com.spring.guideance.tag.domain;
 
-import com.spring.guideance.article.Article;
-import com.spring.guideance.user.User;
+import com.spring.guideance.post.domain.Article;
 import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
 
 @Entity
 @Getter @Setter
