+++ src/main/java/com/spring/guideance/user/domain/Notice.java
@@ -1,10 +1,11 @@
-package com.spring.guideance.user;
+package com.spring.guideance.user.domain;
 
 import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @Getter @Setter
@@ -13,17 +14,12 @@ public class Notice {
     @Column(name = "notice_id")
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
     private int type; // 게시글에 좋아요 발생, 게시글에 댓글 발생, 구독한 태그에 새로운 게시글 등장
 
     private String title;
 
-    private String content;
+    private String contents;
 
-    private LocalDateTime createdAt;
-
-    private boolean checked; // 읽음 여부
+    @OneToMany(mappedBy = "user")
+    private List<UserNotice> userNotices = new ArrayList<>();
 }
